requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.3
    # via requests
idna==3.2
    # via requests
python-frontmatter==1.0.0
    # via -r requirements.in
pyyaml==5.4.1
    # via python-frontmatter
requests==2.26.0
    # via -r requirements.in
urllib3==1.26.6
    # via requests
